requirements.txt
# Core dependencies
openai>=1.12.0
pdf2image>=1.16.0
Pillow>=10.0.0
streamlit>=1.31.0
pandas>=2.2.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0

# Optional dependencies
pytesseract>=0.3.10  # Per eventuali OCR aggiuntivi